// mobile/myapptemp/src/rest/api.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * REST_API
 * BackEndのREST API仕様です
 * OpenAPI spec version: 1.0.0
 */
import axios from 'axios'
import type {
  AxiosRequestConfig,
  AxiosResponse
} from 'axios'
/**
 * Commentオブジェクト
 */
export type MarkerById200Comment = {
  /** ID */
  id?: number;
  /** コメント文字列 */
  comment?: string;
};

/**
 * Markerオブジェクト
 */
export type MarkerById200 = {
  /** ID */
  id: number;
  /** 登録日付(Unixミリ秒) */
  timestamp?: number;
  /** 緯度(latitude) */
  latitude?: number;
  /** 経度(longitude) */
  longitude?: number;
  /** 高度(altitude) */
  altitude?: number;
  /** メートル単位で測定された、場所の不確実性の半径 */
  accuracy?: number;
  /** 高度値の精度 (メートル単位) */
  altitudeAccuracy?: number;
  /** このデバイスの水平方向の移動方向。真北からコンパスを中心に時計回りに度数で測定されます。したがって、北は 0 度、東は 90 度、南は 180 度などです */
  heading?: number;
  /** デバイスの瞬間速度 (メートル/秒) */
  speed?: number;
  /** Commentオブジェクト */
  comment?: MarkerById200Comment;
  /** 画像ID */
  imageId?: number;
};

/**
 * 生成結果ユニークID
 */
export type MarkerCreate200 = {
  /** ユニークID */
  name: string;
};

/**
 * Commentオブジェクト
 */
export type MarkerCreateBodyComment = {
  /** ID */
  id?: number;
  /** コメント文字列 */
  comment?: string;
};

export type MarkerCreateBody = {
  /** 登録日付(Unixミリ秒) */
  timestamp: number;
  /** 緯度(latitude) */
  latitude?: number;
  /** 経度(longitude) */
  longitude?: number;
  /** 高度(altitude) */
  altitude?: number;
  /** メートル単位で測定された、場所の不確実性の半径 */
  accuracy?: number;
  /** 高度値の精度 (メートル単位) */
  altitudeAccuracy?: number;
  /** このデバイスの水平方向の移動方向。真北からコンパスを中心に時計回りに度数で測定されます。したがって、北は 0 度、東は 90 度、南は 180 度などです */
  heading?: number;
  /** デバイスの瞬間速度 (メートル/秒) */
  speed?: number;
  /** Commentオブジェクト */
  comment?: MarkerCreateBodyComment;
  /** 画像ID */
  imageId?: number;
};





  /**
 * Marker登録APIです
 * @summary Marker登録API
 */
export const markerCreate = <TData = AxiosResponse<MarkerCreate200>>(
    markerCreateBody: MarkerCreateBody, options?: AxiosRequestConfig
 ): Promise<TData> => {
    return axios.post(
      `/marker/create.json`,
      markerCreateBody,options
    );
  }

/**
 * Marker取得API(ID指定)です
 * @summary Marker取得API(ID指定)
 */
export const markerById = <TData = AxiosResponse<MarkerById200>>(
     options?: AxiosRequestConfig
 ): Promise<TData> => {
    return axios.get(
      `/marker/select.json`,options
    );
  }

export type MarkerCreateResult = AxiosResponse<MarkerCreate200>
export type MarkerByIdResult = AxiosResponse<MarkerById200>